Fix right movements on map

+++ src/game/fleet/fleet.rs
@@ -45,9 +45,9 @@ pub struct FleetTravelData {
 impl Fleet {
     fn check_travel_destination(&self, origin_coords: Coordinates, dest_coords: Coordinates) -> Result<()> {
         if  dest_coords.x > origin_coords.x + 1 ||
-            dest_coords.x < origin_coords.x - 1 ||
+            (dest_coords.x as i8) < (origin_coords.x as i8) - 1 ||
             dest_coords.y > origin_coords.y + 1 ||
-            dest_coords.y < origin_coords.y - 1 {
+            (dest_coords.y as i8) < (origin_coords.y as i8) - 1 {
                 return Err(InternalError::FleetInvalidDestination)?;
         }
         Ok(())
